Change attack speed logic and add wave data

+++ Assets/Scripts/Data/Character/PlayerStat.cs
@@ -5,7 +5,7 @@ public class PlayerStat : ScriptableObject
 {
     [field: SerializeField, Range(1, 100f)] public float Health = 100;
     [field: SerializeField, Range(1, 100f)] public float MaxHealth = 100;
-    [field: SerializeField, Range(0.001f, 10f)] public float BonusAttackSpeed { get; set; }
+    [field: SerializeField, Range(0.001f, 10f)] public int BonusAttackSpeed { get; set; }
     [field: SerializeField, Range(0.001f, 10f)] public float Damage = 10;
     [field: SerializeField, Range(1, 5f)] public float Speed = 3;
